feat: update to support pending block
https://github.com/ethereum/go-ethereum/pull/23033
https://github.com/ethereum/go-ethereum/pull/24949
https://github.com/ethereum/go-ethereum/pull/25442
https://github.com/ethereum/go-ethereum/pull/25922
https://github.com/ethereum/go-ethereum/pull/26922
https://github.com/ethereum/go-ethereum/pull/27135
https://github.com/ethereum/go-ethereum/pull/28623

+++ api/api_public_klay.go
@@ -78,7 +78,7 @@ type FeeHistoryResult struct {
 
 // FeeHistory returns data relevant for fee estimation based on the specified range of blocks.
 func (s *PublicKaiaAPI) FeeHistory(ctx context.Context, blockCount DecimalOrHex, lastBlock rpc.BlockNumber, rewardPercentiles []float64) (*FeeHistoryResult, error) {
-	oldest, reward, baseFee, gasUsed, err := s.b.FeeHistory(ctx, int(blockCount), lastBlock, rewardPercentiles)
+	oldest, reward, baseFee, gasUsed, err := s.b.FeeHistory(ctx, uint64(blockCount), lastBlock, rewardPercentiles)
 	if err != nil {
 		return nil, err
 	}
